# Covid105.Registration/Model/Patient.cs
using System;

namespace Covid105.Registration.Model
{
    public class Patient
    {
        public int Id { get; set; }
        public string FirstName { get; set; }
        public string LastName { get; set; }
        public int? Age { get; set; }
        public DateTime TestDate { get; set; }
    }
}